refactor(shop): rename shop-members API path + envelope to org-members (P4c)
Phase 4c of the OrgMembership terminology migration. Wire-contract change,
consumers + tests moved in the same commit:

- Route path /api/shop/members -> /api/shop/org-members (git mv).
- Response envelope key members -> orgMembers (camelCase). Path is kebab
  (URL convention) but the JSON key is camelCase so consumers read
  .orgMembers; the divergence is intentional and noted in registry.ts.
- Bag key `Person` unchanged (it drives field-strip, unrelated to "member").
- Consumer ToolManagementPanel.tsx: fetch URL + .orgMembers read/fallback.
- Tests: strip test (renamed file), shopAPI.integration, panel fetch-mocks.

Stacks on P4b (claude/fervent-noether-0b7a8b).
