+++ checkin-app/src/app/api/shop/org-members/route.ts
@@ -4,7 +4,7 @@ import { ACTIVE_ORG_MEMBER_PERSON_WHERE } from "@/lib/orgMembership";
 
 export const dynamic = 'force-dynamic';
 
-export const GET = handler('GET /api/shop/members', async () => {
+export const GET = handler('GET /api/shop/org-members', async () => {
     // Select email too: admins/board are granted everyones:pii and see it. For a
     // certifier the view holds only member+public, so the stripper drops email
     // and keeps name. Field-stripping — not the query — decides who sees what.
